Opening main window after login
- Now the main window opens after successful login;
- Configuring the appropriate window icon for each os;
- Adding spacing for the login button
- Formatting the code
